adjust for line and load transient 1. ripple : adjust for the line transient testing 2. main: add other function index 3. scope: add line transient settings 4. inst: add LPS505 testing 5. BK9141 change the simulation mode for error at line transient

+++ Power_BK9141.py
@@ -472,7 +472,7 @@ class Power_BK9141(GInst):
             time.sleep(wait_time)
             # finished getting back to the initial state
         else:
-            v_res_temp = int(v_res_temp)
+            v_res_temp = float(v_res_temp)
             v_res_temp = v_res_temp + 1
         # need to return the channel after the calibration is finished
